Refactor step position data type and update API routes
- Changed the data type of position coordinates (X, Y) from int to float64 across models, DTOs, and database schema to support more precise positioning.
- Updated API routes to include workflow ID in the step listing and retrieval endpoints for better context and organization.
- Added a migration script to alter the database schema for the steps table to accommodate the new float64 data type for position coordinates.

+++ internal/infrastructure/persistence/read/step_query_repository.go
@@ -32,8 +32,8 @@ type stepRow struct {
 	StepIndex      string
 	ExecutionOrder int
 	TreeIndex      int
-	PositionX      int
-	PositionY      int
+	PositionX      float64
+	PositionY      float64
 	Status         string
 	CreatedAt      time.Time
 	UpdatedAt      time.Time
